added error output for TypeError, comments
