+++ include/color_console/color_console.py
@@ -4,19 +4,18 @@ from colorama import Fore, Style
 def error_answer(answer: str) -> str:
     return Fore.RED + answer + Style.RESET_ALL
 
+
 # print valid bot answer
-
-
 def print_valid(answer: str) -> None:
     print(Fore.GREEN + answer + Style.RESET_ALL)
 
+
 # print valid bot answer
-
-
 def print_error(answer: str) -> None:
     print(error_answer(answer))
 
 
+# wrapper for input
 def colored_input(request: str) -> str:
     # add color for request
     print(colored_request + request, end='')
