+++ include/helpers.py
@@ -1,6 +1,7 @@
 import re
 
 
+# class for store and edit records
 class Field:
     def __init__(self, value: str):
         self.value = value
@@ -13,5 +14,7 @@ class Name(Field):
     pass
 
 
+# validate phone number
+# required 10 digits
 def validate(value: str) -> re.Match | None:
     return re.fullmatch(r'\d{10}', value) is not None
